test-automation/LmsApi/requirement.txt
async-generator==1.10
attrs==22.2.0
azure-core==1.26.3
azure-identity==1.12.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
confluent-kafka==2.0.2
cryptography==39.0.1
dnspython==2.3.0
exceptiongroup==1.1.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
google-api-core==2.11.0
google-auth==2.16.1
googleapis-common-protos==1.58.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
locust-plugins==3.0.2
lxml==4.9.2
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
msgpack==1.0.4
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
outcome==1.2.0
paho-mqtt==1.6.1
playwright==1.30.0
portalocker==2.7.0
protobuf==4.22.0
psutil==5.9.4
psycogreen==1.0.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==9.0.4
PyJWT==2.6.0
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
rsa==4.9
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
unicorn==2.0.1.post1
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.14.0
zope.event==4.6
zope.interface==5.5.2
